changed wording edit selected
This is an attempted resolution to issue#160. This update changes the "Edit Selected" button to say "Mass Edit" while
the user is not in list view, and changes it to say "Edit Selected"
while they are in list view. In addition, pressing the button while not
in list view will switch the user to list view instead of showing a
dialog.

I am not sure about my implementation for changing the text of the
button. An alternative might be to put some javascript into the pug
file, to be evaluated on the page as the user switches views. This might
be preferable as it keeps all display related code in the pug file
(cube_list.pug) rather than in the javascript.

I have also changed the name for the DOM element "editSelected" to
"massEdit" to reflect it's new default value.