Fix spelling of cookie and lengthen expiration. (#197)
app.js misspelled cookie, and so the secure attribute was not actually being set. Additionally, the maximum age was not being set, and so logouts occur frequently. On the current site, this often leads to putting in a bunch of changes and then having them disappear because the "save" button just leads to a re-login.