add search field and query parsing
This update adds a search field to the filter form, and adds logic for
parsing syntax similar to a certain popular online tool that starts with
S and rhymes with try mall. Currently supported for parsing are all of
the words/terms defined in the map object added to editcube.js.

The next commit will handle the logic of actually filtering the cube
based on the filters object.