Scryfall Search Syntax (#313)
* add search field and query parsing

This update adds a search field to the filter form, and adds logic for
parsing syntax similar to a certain popular online tool that starts with
S and rhymes with try mall. Currently supported for parsing are all of
the words/terms defined in the map object added to editcube.js.

The next commit will handle the logic of actually filtering the cube
based on the filters object.

* working recursion with parentheses

Currently OR operator is working as long as user supplies correct
parentheses. Parser only fails to respect operator precedence for
statements like 'a or b c', that have multiple interpretations. Feel
free to change or edit.

* switch to tokenizer + parser

This commit adds a tokenizer, which transforms a string to a list of
only tokens (objects that have an operand, category, not value, and
argument), single parentheses, or the word 'or'. Next commit will bring
the tree building!

* generateTree completed

Now the parser correctly generates a tree with filters and correct
and/or logic. I haven't tested it super extensively, but it's doing very
well. The next update will filter through the tokenized string in an
attempt to make sure it is valid.

* fix tokenizeInput and add verifier

Tokens are now verified before they are processed into a tree, at the
very least this should ensure that parseToken never returns errors. The
last thing to do is verify the entire token string. I haven't figured
out a string that can break the current checks, but we'll see what
happens. I'm sure it's possible.

* name filtering added, UI beginning

The buttons are in place for the UI, a modal for the advanced search
still needs to be added, but reset filters works. Currently, filters are
implemented, and name filtering works. The next commit will add most of
the other filters.

* added oracle text and color searching

Also fixed a bug in editcube, but I've forgotten what it was, I'm so
sorry.

* fixed arg setting

Very simple change, just accidentally used operand instead of
token.operand

* all sorts added and minor bugfixes

Currently everything listed in editcube.js object categoryMap is
implemented and working(?). I've done very little testing, but I think
this is ready for testing on the dev branch.