tokenize bugfix and sort bugfix
Fixed two bugs. In editcube, the tokenize fucntion wouldn't properly
return the value of the yet to be filtered text, instead truncating it
at the next occurence of token.arg. This was particularly problematic
for short/common args like "r" for the color red.

In SortFilter, I forgot to add an argument to a function call, and it
has been added now.