handle 'and' in filter
Currently the term and is implied by the lack of the term OR, and as
such has no actual meaning in the grammar. To that end, if the parser
encounters the word 'and' on it's own, it just ignores it. This has some
slightly undesirable functionality in that searches that incorrectly use
the word 'and' but are otherwise correct, will still work as though the
'and' was not there. i.e. 'and t:creature and t:artifact and' is exactly
equivalent to 't:artifact t:creature' once parsed.