Filter Search Functionality and Correctness Update (#331)
* bugfixes and functionality for filter

This commit accomplishes many things:
1) Move all token parsing logic into token parsing, before the parser
   would pass raw user strings to the filter logic, and it would fail
there if it was going to fail. Now all arguments are checked and put
into a form the filter can use without further processing.
2) Argument checking, all arguments for fields with only specifically
   formatted arguments allowed are now checked at the time of filter
generation. An invalid filter string error is supplied to the user if
their parsed filter fails any of these checks.
3) Implemented support for shard/wedge/guild names in color and color
   identity searches
4) fixed color searches to use the card's actual color, and
   color_identity to use either the color_identity, or the cube owner's
overridden color identity if available.

* handle 'and' in filter

Currently the term and is implied by the lack of the term OR, and as
such has no actual meaning in the grammar. To that end, if the parser
encounters the word 'and' on it's own, it just ignores it. This has some
slightly undesirable functionality in that searches that incorrectly use
the word 'and' but are otherwise correct, will still work as though the
'and' was not there. i.e. 'and t:creature and t:artifact and' is exactly
equivalent to 't:artifact t:creature' once parsed.

* removed extra log statement