Implement filter storage in URL (#346)
When a correct filter Tree is generated from a scryfall input, the
result is saved to the url, not disrupting any existing parameters. If
reset filters is pressed, the url is reset, and if a page is loaded from
a URL containing a filter arugment, that filter is applied to the cube.