Fix: #1690 Fix description reverting to last saved
There was a conditional in the TextEntry.value for the Cube description part of
the form that was passing the `cube.description` to the input if the current
`cube.raw_desc` was falsy.

`cube.raw_desc && cube.raw_desc !== 'undefined' ? cube.raw_desc : cube.description`

Becuase `''` is falsy in javascript, this meant that
deleting the last character of your description would invert the conditional and
put `cube.description` into the input instead of `cube.raw_desc`.

I replaced the truthiness check `cube.raw_desc &&` with an object property check
`Object.prototype.hasOwnProperty(cube, 'raw_desc')` and pulled the whole ternary
out into a function so it could be better documented.

I spoke with @phulin, and was informed that the check against the string literal
`'undefined'` was deliberate, so I left it in.