feat(contributing) streamling first time setup of application
this PR adds more detail to README about key application concepts as well as
scripting the initial setup of running a local instance of CubeCobra.

Running a local instance of CubeCobra is facilitated throug the use of LocalStack,
an AWS emulation that can be run locally. This allows for creation of s3 buckets,
dynamodb tables and cloudwatch log indexes without having an AWS account.