Ensure the addedTmsp is a Number before creating a Date from it, to fix duplicate Date Added labels.
The Date() constructor expects strings to be date/time formatted strings
such as 2024/01/08 or 08/01/2024 etc. When the string value is a unix
timestamp it does not parse correctly and returns an invalid date.

When invalid dates slip into the sort labels (getLabelsRaw) it breaks logic
that consolidates the labels into a sorted set of MM/DD/YYYY strings.