Check if card is already in the removal set before adding to it.
Fixes adding the same card multiple times to the removal list (based on
the cube card index). If that occurred and was committed, extra cards
will be removed because the removals are done in decreasing index
order, so after the targeted card is removed whatever other card now
resides at the index gets removed by the 2nd instance of the removal in
the change set.