Allow developers to run the app in production mode with docker.
Normally requests to through webpack dev server listening on 8080. In
docker those are passed through to node listening on 5000 within the container, meaning
the server only has to listen for connections on 127.0.0.1.

In production mode there is no webpack dev server and we can direct the
8080 requests to 5000 with docker. But since the incoming connection is
a non-local IP, we have to allow node to accept connections from any IP.