Make the Group Modal and Package Modal card lists scrollable.
For example in Group Modal of Cube list, when you open a group like all
green creatures the list can be very long. Though with scrollable modals
themselves that is OK, it is even better UX for the list to be
scrollable so that actions like the Purchase buttons are visible right
away.