Stop click event propagation in the ConfirmDeleteModal, to prevent parent clock actions from occur after the delete or cancel.
For example the DeckPreview is a row that has a click action to redirect to the deck.
Within that it has a button that opens the Confirm delete modal. Without
the propagation stopping, the click on Delete also triggered the
redirect action on the row; and depending on timing the request to the
server to Delete may not have completed.