Fix errors caused by pending picks and failing bot predictions.
Most likely the failure is the client side timeout.

When the predictions failed and there is a pending pick, it proceeded to
the pick logic but the predictions in state were for the previous packs.
The predicted oracle ids wouldn't map to the current packs, tries to
access index -1 of the pack, gets undefined. When the state is saved
into localstorage that becomes null.

Then finally when the draft finishes, the server side validation rejects
the request immediately because it expects all picks to be numbers not
null.

Two different mitigations applied:
1) Don't proceed with pending pick if there are prediction errors or
   retries in process
2) If the oracle ID cannot be mapped in the pack, randomly choose the
   index